refactor: move load_tokenizer, load_model, get_dataset, and  get_loaders functions to utils directory for reusability.
